'Serialize' MambuStruct.
Previous way of 'serializing' only 'serializes' the fields at the
beginning of the life of the MambuStruct (right after loading it from
Mambu).

The greate problem with this approach is that, when already loaded and
then something changes, this would not be reflected on the
'serialization'.

Another problem is that only the fields that are taken from Mambu and
that are numbers or strings or datetimes are the ones to be
'serialized', but if you add another MambuStruct as a member of the
MambuStruct, this one wouldn't be 'serialized'.

serializeStruct method is added to 'serialize' MambuStructs.

Helper method serializeFields is added, an static method in MambuStruct,
just to iterate through its fields. If a field is in turn a MambuStruct,
it calls serializeStruct recursively. Each field in the iterable calls
serializeFields recursively. In any other case, it ends the recursion by
'serializing' the unicode()

The previous method, a serial attribute, gets deprecated.