Some arguments for the method setCustomField were removed
There is a field in each customfield that defines what type of field
is, like number, selection, string, user_link, etc., so the argument
fieldtype is not necessary.
The message for the exception was changed to english.
If the datatype of the customfield is not USER_LINK the original
value is setted for the new field.