Allow the use of grouped custom fields from Mambu.
Grouped custom fields allow users to add the same custom field a lot
of times on the entity. This affects on that the several customfields
from grouped custom fields on MambuPy have the same name.

A mechanism was implemented to index each name from grouped custom
fields.

setCustomField() method was changed too so that custom fields are
searched using the name, not the custom field ID.