MambuStruct add 'email' to convertDict2Attrs() constantFields
constantFields tells convertDict2Attrs which fields not to try and
convert to an obvious type (everything comes as strings, this method
tries first to convert numbers as ints, dates as datetimes etc.)

email should be left as a string. That should be obvious but also
users are allowed to input anything on an email field on Mambu, so
this commit works as a safeguard in case someone types, say a number,
anticipating that this method would try to convert it to an int, while
in fact it should stay as string.