API V2 mambu update.
Updates the data of certain entity using the update endpoint.

Example of Mambu endpoint documentation for Loan Accounts:
https://api.mambu.com/?python#loan-accounts-update

Mambu in itself won't allow update of some fields on several
cases (for example on account status that imply disbursement or
closing of the account). In those cases MambuPy re-raises the error
given by Mambu.