APIv2 assign entity objects method.
method `MambuStruct._assignEntObjs` loops _entities list looking for
encodedkeys in the MambuPy entity object to get from Mambu.

_entities list conforms of tuples stating the key to look, the class
to get and instantiate, and the name of the property that will hold
the resulting object

get_entities parameter controls if this happens automatically or
not. BEWARE that defaulting to True may imply a lot of requests to
Mambu and a lot of time spent until every entity is got from Mambu

Value Objects may also have an _entities list, because some VOs may
hold keys to Mambu entities too.