feature: restito adapter for v1 to v2 interface compatibility.
MambuStruct inherits from v1's MambuStruct, but has-a wrapped2
property, with an instance of a v2 MambuEntity.

It also has special __getitem__, __setitem__, __getattribute__ and
__setattr__ methods that use the wrapped2 property to access v2
properties, and defaults to v1 in case of error.

Works great for methods of v1 not yet supported by the adapter.

TODO: still a lot of refinement for properties of v1.