feature: logging in MambuPy
A pretty elemental logging facility for MambuPy using logging library.

It is by default turned off, but propagates to the 'mambupy' logger

If, on your application, you configure a mambupy logger, it will
'catch' what MambuPy logs here. Otherwise it will remain silent.

A default yaml configuration and python script for logging is
included in the utils module.