feat: v2 PATCH operations assume fields even if non-existent.
BEFORE: PATCH operations assumed that if a field couldn't be find in
the original entity definition, nor in any of its custom fields or
customfields-inside-a-set, then the field is invalid, so no PATCH
could be done there.

What about if the entity didn't had some field but could have it? Lets
say, an assigned Branch. The assignedBranchKey would be inexistent on
the original entity, and you would not be able to PATH the entity for
assigning it to a Branch.

AFTER: PATCH assume that the field may exist by default in the root of
the entity if not found on any other place.

So now, you can assign the entity to a branch on our previous example.

You alse may try to PATCH an invalid field to Mambu. Which will
non-graciouslly break your code with an error from Mambu telling you
that you may be on some kind of substance inventing fields by
yourself.