fix: v1 retries do not consider 500 errors.
retry on 429, 502, 503 & 504. But 500, do not retry.

Usually Mambu responds a 500 with error messages from conditions not
originating on errors from Mambu, such as payments not viable to be
made because the loan account has a wrong status, etc. (Maybe that
should be a 412 or other, but Mambu rules here, and sends 500 on such
situations).